fix: create crashing on server - move some mixins to client-only
